Laboratory 4 updated

+++ Laboratory4/src/main/java/moldovan/christine/lab4/ex1/Circle.java
@@ -19,6 +19,6 @@ public class Circle {
     }
 
     public double getArea() {
-        return (double) Math.PI * (radius * radius);
+        return Math.PI * (radius * radius);
     }
 }
